Add files via upload
The Hong Leong Bank event with the "Gatsby in the Orient" theme was a spectacular affair that seamlessly blended the opulence of the Gatsby era with the allure of the Orient. The meticulously designed photo opportunity, stage, and trophy all played a pivotal role in bringing this theme to life. The photo op exuded the elegance of the Gatsby movie while incorporating Oriental elements, featuring thematic furniture, Hand fans, and decor inspired by both cultures. The trophy, in gold and black, not only symbolized achievement but also encapsulated the grandeur and sophistication associated with the Gatsby era, making it a fitting recognition for the occasion. Together, these design elements transformed the event into a captivating fusion of two worlds, leaving a lasting impression on attendees.